Добавил 98% покрытия тестами

+++ app/main.py
@@ -2,23 +2,10 @@ from contextlib import asynccontextmanager
 from typing import Annotated
 
 from fastapi import Depends, FastAPI, HTTPException, Query
-from pydantic_settings import BaseSettings, SettingsConfigDict
-from sqlmodel import Field, Session, SQLModel, create_engine, select
+from sqlmodel import Field, SQLModel, select
 
-from register import reg
-
-
-class Setting(BaseSettings):
-    key: str
-    db: str
-    postgres_password: str
-
-    model_config = SettingsConfigDict(env_file=".env")
-
-
-setting = Setting()  # type: ignore
-
-engine = create_engine(setting.db)
+from app.database import SessionDep, engine
+from app.register import Librarian, User, get_current_user, reg
 
 
 class BookBase(SQLModel):
@@ -41,14 +28,6 @@ def create_db():
     SQLModel.metadata.create_all(engine)
 
 
-def get_session():
-    with Session(engine) as session:
-        yield session
-
-
-SessionDep = Annotated[Session, Depends(get_session)]
-
-
 @asynccontextmanager
 async def lifespan(app: FastAPI):
     create_db()
@@ -62,6 +41,7 @@ app.include_router(reg)
 
 @app.get("/books")
 def books(
+    user: Annotated[User, Depends(get_current_user)],
     session: SessionDep,
     offset: int = 0,
     limit: Annotated[int, Query(le=100)] = 100,
@@ -70,18 +50,24 @@ def books(
     return books
 
 
-@app.get("/book/{book_id}")
-def one_book(book_id: int, session: SessionDep):
+@app.get("/books/{book_id}")
+def one_book(
+    book_id: int, user: Annotated[User, Depends(get_current_user)], session: SessionDep
+):
     book = session.get(Book, book_id)
     if not book:
-        raise HTTPException(
-            status_code=404, detail="Enter name athere book this book not in db"
-        )
+        raise HTTPException(status_code=404, detail="Book not found")
     return book
 
 
-@app.post("/book/add/")
-def add_book(book: BookBase, session: SessionDep):
+@app.post("/books")
+def add_book(
+    book: BookBase,
+    librarian: Annotated[Librarian, Depends(get_current_user)],
+    session: SessionDep,
+):
+    if not librarian.librarian:
+        raise HTTPException(status_code=403, detail="Permission denied")
     db_book = Book.model_validate(book)
     session.add(db_book)
     session.commit()
@@ -89,25 +75,34 @@ def add_book(book: BookBase, session: SessionDep):
     return db_book
 
 
-@app.delete("/book/{book_id}")
-def delete_book(book_id: int, session: SessionDep):
+@app.delete("/books/{book_id}")
+def delete_book(
+    book_id: int,
+    librarian: Annotated[Librarian, Depends(get_current_user)],
+    session: SessionDep,
+):
+    if not librarian.librarian:
+        raise HTTPException(status_code=403, detail="Permission denied")
     book = session.get(Book, book_id)
     if not book:
-        raise HTTPException(
-            status_code=404, detail="Enter name another book this book not in db"
-        )
+        raise HTTPException(status_code=404, detail="Book not found")
     session.delete(book)
     session.commit()
     return {"Book delete": True}
 
 
 @app.patch("/book/{book_id}")
-def patch_book(book_id: int, book: BookUpdate, session: SessionDep):
+def patch_book(
+    book_id: int,
+    book: BookUpdate,
+    librarian: Annotated[Librarian, Depends(get_current_user)],
+    session: SessionDep,
+):
+    if not librarian.librarian:
+        raise HTTPException(status_code=403, detail="Permission denied")
     book_db = session.get(Book, book_id)
     if not book_db:
-        raise HTTPException(
-            status_code=404, detail="Enter name another book this book not in db"
-        )
+        raise HTTPException(status_code=404, detail="Book not found")
     book_data = book.model_dump(exclude_unset=True)
     book_db.sqlmodel_update(book_data)
     session.add(book_db)
